Add files via upload
For safe and secure lending experience, it's important to analyze the past data. In this project, 
you have to build a deep learning model to predict the chance of default for future loans using 
the historical data. As you will see, this dataset is highly imbalanced and includes a lot of 
features that make this problem more challenging.